otel/collector-load-test/.local-config/python/requirements.txt
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
googleapis-common-protos==1.61.0
grpcio==1.59.3
idna==3.6
importlib-metadata==6.11.0
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
protobuf==4.25.1
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.1.0
wrapt==1.16.0
zipp==3.17.0
